pyproject.toml
[project]
name = "krypt"
version = "0.1.0"
description = "Krypt is an end-to-end encrypted messaging app with a TUI."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.1",
    "asyncpg>=0.31.0",
    "fastapi[standard]>=0.128.0",
    "psycopg2>=2.9.11",
    "pwdlib[argon2]>=0.3.0",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
    "sqlalchemy>=2.0.45",
]

[tool.setuptools.packages.find]
where = ["src"]
